requirements.txt
mfrc522-python==0.0.7
numpy==2.1.1
pandas==2.2.3
pyembedded==3.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
pytz==2024.2
shiboken6==6.7.2
six==1.16.0
tzdata==2024.2